// src/app/models/test-yesavage.ts
export class TestYesavage {
    //ey_id: String;
    ef_id: String;
    ey_p1_satisfecho: String;
    ey_p2_actividades: String;
    ey_p3_vacio: String;
    ey_p4_aburrido: String;
    ey_p5_animo: String;
    ey_p6_preocupado: String;
    ey_p7_feliz: String;
    ey_p8_desamparado: String;
    ey_p9_cosas: String;
    ey_p10_memoria: String;
    ey_p11_estar_vivo: String;
    ey_p12_inutil_despreciable: String;
    ey_p13_energia: String;
    ey_p14_esperanza_actual: String;
    ey_p15_cree_mejor: String;
    ey_tiempo_inicial: String;
    ey_tiempo_final: String;
    ey_tiempo_total: String;
    ey_estado: Boolean;
    ey_puntaje_total: Number;
}